+++ packages/auth-passport/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @keystone-alpha/auth-passport
 
+## 4.1.4
+
+### Patch Changes
+
+- Updated dependencies [[`6d7d0df0`](https://github.com/keystonejs/keystone-5/commit/6d7d0df0515c3aa21c7d24db17919ddbb5701ce9)]:
+  - @keystone-alpha/fields@14.0.0
+
 ## 4.1.3
 
 - Updated dependencies [4e6a574d](https://github.com/keystonejs/keystone-5/commit/4e6a574d):