+++ test-projects/client-validation/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @keystone-alpha/cypress-project-client-validation
 
+## 1.0.16
+
+### Patch Changes
+
+- Updated dependencies [[`6d7d0df0`](https://github.com/keystonejs/keystone-5/commit/6d7d0df0515c3aa21c7d24db17919ddbb5701ce9)]:
+  - @keystone-alpha/adapter-mongoose@6.0.0
+  - @keystone-alpha/fields@14.0.0
+  - @keystone-alpha/keystone@16.0.0
+  - @keystone-alpha/app-admin-ui@5.10.2
+
 ## 1.0.15
 
 - Updated dependencies [4e6a574d](https://github.com/keystonejs/keystone-5/commit/4e6a574d):